Implement Phase 2: Domain Data Layer (`Row` + `Index`)
This commit implements Phase 2 of the 5.1 Redo plan by adding `RID int64`
and `Deleted bool` to the `Row` struct and updating the `Index` struct
to store a list of `int64` rather than `int`. The serialization logic was
also updated to account for the new 9 byte offset at the start.

Additionally, the `5.1plan.md` file was updated with Phase 2 checked off
and notes added for future reference.

+++ internal/domain/data/index.go
@@ -3,6 +3,6 @@ package data
 // Index is an in-memory index on a single column
 type Index struct {
 	Column string
-	Data   map[interface{}][]int // value → row positions
+	Data   map[interface{}][]int64  // value → RIDs (was []int positions)
 	Unique bool
 }
